Modify project template and refactor code
