+++ packages/base_package/package.xml
@@ -1,0 +1,19 @@
+<package>
+  <name>base_package</name>
+  <version>0.1.0</version>
+  <description>
+    Base class and functionality package
+  </description>
+  <maintainer email="giokepa4@gmail.com">Giorgi Tsartsidze</maintainer>
+  <license>None</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>duckietown_msgs</build_depend>
+  <build_depend>sensor_msgs</build_depend>
+  <build_depend>rospy</build_depend>
+
+
+  <run_depend>duckietown_msgs</run_depend>
+  <run_depend>sensor_msgs</run_depend>
+  <run_depend>rospy</run_depend>
+</package>
