+++ packages/image_processing_package/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(my_package)
+project(image_processing_package)
 
 find_package(catkin REQUIRED COMPONENTS
 	rospy
