+++ packages/image_processing_package/package.xml
@@ -1,8 +1,8 @@
 <package>
-  <name>my_package</name>
+  <name>image_processing_package</name>
   <version>0.1.0</version>
   <description>
-    Package of April Tagging
+    Image processing package, which has apriltagger and its corresponding node
   </description>
   <maintainer email="giokepa4@gmail.com">Giorgi Tsartsidze</maintainer>
   <license>None</license>
